Autoencoder architecture modifief
Now the imput is a tensor (t,1,f) where:
t is the length of the series
1 is because we consider only one device at time
f is the number of features we provide (derivations ).
The imput tensor is simply filtered by a Gaussian Kernel with the stdev proporitonal to the variance of the temperature of the device we are considering.

The output has obviusly the same shape of the input and the data we want to map is the same as the input but now we pick only the seasonal component of the time serie